+++ src/jacobi.jl
@@ -1,8 +1,18 @@
+"""
+$(TYPEDEF)
+
+Jacobi preconditoner
+"""
 struct JacobiPreconditioner{Tv, Ti} <: AbstractExtendablePreconditioner{Tv,Ti}
     extmatrix::ExtendableSparseMatrix{Tv,Ti}
     invdiag::Array{Tv,1}
 end
 
+"""
+$(SIGNATURES)
+
+Update Jacobi preconditoner
+"""
 function update!(precon::JacobiPreconditioner)
     cscmatrix=precon.extmatrix.cscmatrix
     invdiag=precon.invdiag
@@ -13,6 +23,11 @@ function update!(precon::JacobiPreconditioner)
     precon
 end
 
+"""
+$(SIGNATURES)
+
+Construct Jacobi preconditoner
+"""
 function JacobiPreconditioner(extmatrix::ExtendableSparseMatrix{Tv,Ti}) where {Tv,Ti}
     @assert size(extmatrix,1)==size(extmatrix,2)
     flush!(extmatrix)
@@ -21,9 +36,19 @@ function JacobiPreconditioner(extmatrix::ExtendableSparseMatrix{Tv,Ti}) where {T
     update!(precon)
 end
 
+"""
+$(SIGNATURES)
+
+Construct Jacobi preconditoner
+"""
 JacobiPreconditioner(cscmatrix::SparseMatrixCSC{Tv,Ti}) where {Tv,Ti}=JacobiPreconditioner(ExtendableSparseMatrix(cscmatrix))
 
 
+"""
+$(SIGNATURES)
+
+Solve  Jacobi preconditioning system
+"""
 function  LinearAlgebra.ldiv!(u::AbstractArray{T,1} where T, precon::JacobiPreconditioner, v::AbstractArray{T,1} where T)
     invdiag=precon.invdiag
     n=length(invdiag)
@@ -32,6 +57,11 @@ function  LinearAlgebra.ldiv!(u::AbstractArray{T,1} where T, precon::JacobiPreco
     end
 end
 
+"""
+$(SIGNATURES)
+
+Inplace solve  Jacobi preconditioning system
+"""
 function LinearAlgebra.ldiv!(precon::JacobiPreconditioner, v::AbstractArray{T,1} where T)
     ldiv!(v, precon, v)
 end