+++ src/parallel_jacobi.jl
@@ -1,8 +1,18 @@
+"""
+$(TYPEDEF)
+
+Parallel Jacobi preconditioner
+"""
 struct ParallelJacobiPreconditioner{Tv, Ti} <: AbstractExtendablePreconditioner{Tv,Ti}
     extmatrix::ExtendableSparseMatrix{Tv,Ti}
     invdiag::Array{Tv,1}
 end
 
+"""
+$(SIGNATURES)
+
+Parallel Jacobi preconditioner update
+"""
 function update!(precon::ParallelJacobiPreconditioner)
     cscmatrix=precon.extmatrix.cscmatrix
     invdiag=precon.invdiag
@@ -13,6 +23,11 @@ function update!(precon::ParallelJacobiPreconditioner)
     precon
 end
 
+"""
+$(SIGNATURES)
+
+Parallel Jacobi preconditioner construtor
+"""
 function ParallelJacobiPreconditioner(extmatrix::ExtendableSparseMatrix{Tv,Ti}) where {Tv,Ti}
     @assert size(extmatrix,1)==size(extmatrix,2)
     flush!(extmatrix)
@@ -21,9 +36,19 @@ function ParallelJacobiPreconditioner(extmatrix::ExtendableSparseMatrix{Tv,Ti}) 
     update!(precon)
 end
 
+"""
+$(SIGNATURES)
+
+Parallel Jacobi preconditioner construtor
+"""
 ParallelJacobiPreconditioner(cscmatrix::SparseMatrixCSC{Tv,Ti}) where {Tv,Ti}=ParallelJacobiPreconditioner(ExtendableSparseMatrix(cscmatrix))
 
 
+"""
+$(SIGNATURES)
+
+Parallel Jacobi preconditioner solve
+"""
 function  LinearAlgebra.ldiv!(u::AbstractArray{T,1} where T, precon::ParallelJacobiPreconditioner, v::AbstractArray{T,1} where T)
     invdiag=precon.invdiag
     n=length(invdiag)
@@ -32,6 +57,11 @@ function  LinearAlgebra.ldiv!(u::AbstractArray{T,1} where T, precon::ParallelJac
     end
 end
 
+"""
+$(SIGNATURES)
+
+Parallel Jacobi preconditioner inplace solve
+"""
 function LinearAlgebra.ldiv!(precon::ParallelJacobiPreconditioner, v::AbstractArray{T,1} where T)
     ldiv!(v, precon, v)
 end